Add files via upload
In this project, we performed data wrangling on the Titanic dataset using Seaborn and Pandas. We handled missing values, created new features, removed unnecessary columns, and saved the cleaned dataset. The dataset is now ready for further analysis or machine learning tasks. Data wrangling is a crucial step in any data science project, as it ensures the data is clean, structured, and meaningful.